backend/pyproject.toml
[project]
name = "memory-break-orchestrator"
version = "0.1.0"
description = "Memory-Break Orchestrator for AI agent evaluation following VIBE architecture"
authors = [{name = "Memory-Break Team", email = "team@example.com"}]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.13"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlmodel>=0.0.14",
    "alembic>=1.12.1",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.7",
    "redis>=5.0.1",
    "rq>=1.15.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "httpx>=0.25.2",
    "pexpect>=4.8.0",
    "GitPython>=3.1.40",
    "jinja2>=3.1.2",
    "websockets>=12.0",
    "aiofiles>=23.2.1",
    "aiohttp>=3.9.0",
    "psutil>=5.9.0",
    "anthropic>=0.72.0",
    "google-generativeai>=0.3.2",
    "openai>=1.3.7",
    "iflow-cli-sdk>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-mock>=3.12.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "ruff>=0.1.6",
    "pre-commit>=3.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "UP", "B", "SIM", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
